Fix spawning circles that overlap initial circles
Could have occurred during the first 3 seconds (default circles lifespan)
when the random new circle position was close enough
to the initial circles in the middle,
and only at non-default resolution of canvas (different than 1080x1920).

Also was causing another hard-to-find issue
‒ the initial circles positions, mistakenly multiplied by resolution scale,
were not erased from the spawn positions list,
which resulted in circles not being able to be created at those positions.

The bug was introduced in the commit 560941f.